Yes/No voting succesfully implemented
feat: YES/NO VOTING
I changed the voting model to include a new type of question, Yes/No. In the process, I also changed the QuestionOption model to include a new type of option, Yes/No. I also changed the serializers and views to handle the new question and option types. I also added a migration to create the new question and option types in the database. #2